Ensured that `ClearItems()` correctly adds attributes as `DeletedAttributes`
When `Remove()` is called, the `RemoveItem()` override correctly adds the removed `AttributeValue.Key` to the `DeletedAttributes` collection, which effectively marks the `AttributeValueCollection` as `IsDirty()`. When calling `Clear()`, however, this did not happen. That's because `Clear()` doesn't call `Remove()`.

To mitigate that, I've added a `ClearItems()` override to `AttributeValueCollection`, and it will add all current `AttributeValue.Key` values to the `DeletedAttributes` collection, before calling `base.ClearItems()`.