Introduced a new `TopicMultiMap` class
The `TopicMultiMap` is a dictionary—or, technically, a `KeyedCollection`—which provides a 1:n relationship between a given collection `key` and a collection of `Topic` objects.

In addition to the basic features of a `KeyedCollection`, it includes convenience overloads which include not only the `key`—as the base methods do—but also the `Topic` as well. As such, it includes e.g. `Contains(key, topic)`, `Add(key, topic)`, and `Remove(key, topic)`. It also has a `Clear()` overload which will clear one of the target collections based on the collection `key`.

Unlike a `Dictionary<TKey, TValue>`, which uses a `KeyValuePair<TKey, TValue>` object, the `TopicMultiMap` introduces its own `KeyValuesPair<TKey, TValue>` class. This therefore operates very similar to a dictionary, but with more intuitive semantics given that the `Values` property is a collection.

These are intended to support an updated `Topic.Relationships` type, though they're not implemented with any relationship-specific semantics, and are thus stored in the `OnTopic.Collections` namespace to denote that they're intended to be multi-purpose.