Introduced a new `ReadOnlyTopicMultiMap`
The `ReadOnlyTopicMultiMap` operates similar to the new `TopicMultiMap`—and, indeed, requires a `TopicMultiMap` object as a source object—but only exposes a read only interface using an `IEnumerable<KeyValuesPair<string, ReadOnlyCollection<Topic>>>`.

This ensures it doesn't have any confusing or unnecessary interface elements associated with e.g. `IDictionary` or `ICollection`, while still allowing users to enumerate and query the collection similar to an actual `TopicMultiMap`. The main difference is that the methods return a `ReadOnlyCollection<Topic>` or `IEnumerable<ReadOnlyCollection<Topic>>` instead of a `Collection<Topic>` or `Collection<KeyValuesPair<string, Collection<Topic>>`, as the `TopicMultiMap` does.

This is intended to act as the base class for the underlying type of `Topic.Relationships`, as we want that to offer the semantics of a collection, while forcing users to update relationships using a controller interface in order to ensure we can enforce business logic such as tracking state and handling recipricol relationships.

Note: This doesn't implement `IDictionary<string, ReadOnlyCollection<Topic>>`, even though it implements several related interface members, as that would force it to return an `IEnumerator<KeyValuePair<string, ReadOnlyCollection<Topic>>`, which introduces confusing semantics for a multi-map (e.g., `collection.Value` vs. `collection.Values`).