Established new `TopicRelationshipMultiMap` from `RelatedTopicCollection`
The legacy `RelatedTopicCollection` has been renamed to `TopicRelationshipMultiMap` and reconfigured to a) derive from the new `ReadOnlyTopicMultiMap` class (eb5c1c3) in order to provide read-only access to relationships, while b) exposing a façade to a private `TopicMultiMap` object for write access.

This ensures that the `TopicRelationshipMultiMap` maintains the semantics of a collection, while enforcing write access through a limited interface which is not only friendlier, but also enforces business logic such as handling reciprocal relationships and state tracking.

Speaking of state tracking, the `IsDirty()` status is now tracked internally within the `TopicRelationshipMultiMap` instead of "polluting" the more general downstream objects, such as the `TopicMultiMap` or the `Collection<Topic>` which it relies upon. This is more consistent with how state tracking is (now) handled elsewhere. This is a bit more involved, but helps centralized the tracking functionality.

Overall, however, this class is now greatly simplified, since much of the functionality is now handled by the underlying `ReadOnlyTopicMultiMap` or the internal `TopicMultiMap`.