Removed legacy artifacts of `IReadOnlyDictionary<>`
Initially, the `ReadOnlyTopicMultiMap` was implemented as an `IReadOnlyDictionary<>`. This didn't work since `IReadOnlyDictionary<>` returns an `IEnumerator` of `KeyValuePair`, which introduces unintuitive semantics for a multimap. To mitigate that, I changed it to an `IEnumerable<KeyValuesPair<string, Collection<Topic>>`.

Given this, the `Values` member is no longer needed, nor do we need to suppress CA1710 (which recommends that classes implementing `IReadOnlyDictionary<>` use the `Dictionary` suffix in their name). Further, we can now return actual `ReadOnlyCollection<>` instances instead of `IEnumerable`. This is useful as it is more consistent with the original interface of the legacy `RelatedTopicCollection`, and will thus prevent some breaking changes.