Updated references to the `Topic.Relationships` class to new interface
The change from `RelatedTopicCollection` (which was a `KeyedCollection<string, NamedTopicCollection>`) to `TopicRelationshipMultiMap` (which is now an `IEnumerable<KeyValuesPair<string, ReadOnlyCollection<Topic>>>`) is broadly consistent in terms of primary entry points, but does change the interface some when dealing with iterators.

This update resolves those breaking changes by e.g., changing the relationship `Name` key to `Key`,  accessing the values from the new `Values` property, and working with `Topic` instead of `Topic.Key` (to account for the fact that it now allows duplicate key names).

These are breaking changes that may need to be applied in implementors as well—though, the hope, is that most implementors will be relying on the higher-level interfaces and, thus, not be affected by this.