Removed legacy `NamedTopicCollection`, merged unit tests
The (clumsily named) `NamedTopicCollection` was used by the legacy `RelatedTopicCollection` (57ff3f9), and is no longer used with the new implementation of the `TopicMultiMap` (d049ea9) and `TopicRelationshipMultiMap` (57ff3f9).

As such, the `NamedTopicCollection` and its related `NamedTopicCollectionTest` unit tests can be deleted. The core functionality of the unit tests, however, remains relevant, and has now been merged into the `RelatedTopicCollectionTest` unit tests, to operate off of the new `TopicRelationshipMultiMap` class.