Introduced new `TopicCollection`
With the legacy `TopicCollection` renamed to the more descriptive `KeyedTopicCollection`, we can now introduce a new `TopicCollection` class which is _not_ keyed, and is thus suitable for cases where we expect topics from across the topic graph to be represented. It's unclear what, if any, additional members we will want to add to this—though having it helps formalize the return types for the few cases where this is necessary, and works off of a more consistent naming convention.