Introduced new `ReadOnlyTopicCollection`
With the legacy `ReadOnlyTopicCollection` renamed to the more descriptive `ReadOnlyKeyedTopicCollection`, and the introduction of a new `TopicCollection` class (9b68bb2), we can now introduce a new `ReadOnlyTopicCollection` class which is _not_ keyed, and is thus suitable for cases where we expect topics from across the topic graph to be represented.

It's unclear what, if any, additional members we will want to add to this—though having it helps formalize the return types for the few cases where this is necessary, and works off of a more consistent naming convention.

Note: This class is not generic as we don't anticipate much use of topic derivatives moving forward—and in the few cases that we do, we expect they'll want to be keyed. For instance, the `ContentTypeDescriptorCollection` and `AttributeDescriptorCollection` both operate off of `Topic` derivatives, and are intended to have unique keys in context, so using the generic `KeyedTopicCollection<T>` is preferred.