Ensured empty constructor is supported
The `Collection<Topic>` constructor accepts an `IList<Topic>` as an optional parameter to prepopulate the collection. But, if that parameter is supplied, it cannot be nullable. As such, in order to keep the parameter nullable, we need to fall back to a new `List<Topic>` instead of passing `null`. Whoops.