Prefer `ReadOnlyTopicCollection` over e.g. `ReadOnlyCollection<Topic>`
Implemented the new `ReadOnlyTopicCollection` to replace instances of where `ReadOnlyCollection<Topic>` had been used. At this point, this doesn't add much functionality, but it's a bit shorter, and offers a more consistent naming convention.

This includes the extension methods previously updated to use `TopicCollection`, since we expect these results to be read-only.