Added mapping support for derived collections
Previously, the collection support of the `TopicMappingService` expected that the collection would be generic, and that's where it would determine the compatible list type.

That works fine if, in fact, the collection type _is_ generic. It fails, however, if the class derives from a generic type. In that case, instead, we need to loop through the available interfaces, identify the `IList<T>` interface, and then get the generic type from _that_ interface, since we won't be able to get it off the actual non-generic type.

This implements that support.