Introduced `referenceTopic` parameter into `ITopicRepository.Load()`
According to the `ITopicRepository.Load()` signature, a topic can be loaded from anywhere within the topic graph. And, indeed, that works just fine assuming one is querying e.g. the `CachedTopicRepository`, which has access to the entire tree.

If one queries a data store—e.g., via the `SqlTopicRepository`—however, that can result in an incomplete topic being returned, since it won't be able to find topic references and relationships which occur outside the `Load()` scope. That can then cause problems if the topic is then saved, as those orphaned references may get overwritten.

(Note: In practice, this usually only affects `Load(topicId, version)` since most applications use the `CachedTopicRepository`. And since relationships and references aren't currently versioned, this doesn't have an impact on `Rollback()` either. Nevertheless, it remains a hole in the design that could yield unexpected bugs, if not data loss.)

As a first step to resolving this, the `ITopicRepository.Load()` signature is extended to accept an optional `referenceTopic`. When present, this prepopulates the list of topics with those already in memory. This ensures that a topic being loaded has access to the rest of the graph, at least insofar as the current application is aware of it.

Further, this optionally allows `ITopicRepository` instances to update existing instances of topics, instead of loading new object references referring to the same entity, which can cause additional problems.

While this commit introduces the extension of the signature, it doesn't implement the overload in the code (except as part of `ITopicRepository` passthroughs), nor does it implement it in any of the concrete implementations, such as `SqlTopicRepository`. Those steps will come in future commits.