Pass `referenceTopic` where available
While the `referenceTopic` isn't yet wired up in any of the concrete `ITopicRepository` implementations—namely `SqlTopicRepository`—the interface now exists in `ITopicRepository` and should be utilized where possible. In this setup, a current topic from the cached topic graph will be passed to `ITopicRepository.Load()` if one is available.

Note: In most cases, we actually expect the implementation to be operating against a `CachedTopicRepository` decorator, in which case this doesn't provide any value. Nevertheless, the OnTopic Library is not aware of what `ITopicRepository` implementation will be used in any application, and should not make any assumptions.

There remain some cases where one isn't available—such as when bootstrapping a call.