Establish new `TopicIndex` collection
A `TopicIndex` is simply a lookup table of `Topic` objects keyed by `Topic.Id`. By default, using e.g., `Topic.Load(topicId)` or even `Topic.FindFirst(t => t.Id == topicId)` is potentially expensive as it requires a recursive search of the topic graph. That's fine for one-time queries. If multiple queries need to be performed, a `TopicIndex` offers better performance by caching the results in an index.

This is primarily intended to support backend infrastructure, though it may be of use for some specialized client implementations.