Introduce `referenceTopic` parameter to `LoadTopicGraph()` extension
The `SqlDataReader.LoadTopicGraph()` extension method is the primary entry point for the `SqlTopicRepository.Load()` method. By introducing this parameter, we ensure that the `referenceTopic` is accounted for to any calls to `SqlTopicRepository.Load()`.

Notably, this initializes the `referenceTopic` by calling `GetRootTopic()` followed by `GetTopicIndex()`, thus prepopulating the topic index used by subsequent methods. This effectively replaces the previous `Dictionary<int, Topic>` used by those methods.

If no `referenceTopic` is passed, or a topic can't be found in the `TopicIndex`, then the code continues to work exactly like it did before: a new `Topic` is created and added to the index. If the `Topic` already exists, however, it is used instead.