Added update logic to the `SqlDataReader` extensions
Currently, `TopicRepositoryBase.Rollback()` loads a new topic via `ITopicRepository.Load()`, then compares the attributes to the existing attributes manually, and updates them accordingly. This is a bit clumsy, and won't will cause problems if we choose to version relationships and references in the future (since reciprocol relationships would point to the temporary topic).

To mitigate that, the logic of the `SqlDataReader` extensions now supports update logic if a `topicReference` is passed in. This allows e.g. the new `Load(topicReference, version)` overload to pass in the current version, and have that instance be updated directly, instead of creating a temporary object and manually merging the values.

This will also lay the groundwork for a proposed `ITopicRepository.Refresh()` method in the future for handling cache updates. (Though that will be a separate project independent of this branch.)

In most cases, this was already supported implicitly since calls were made to e.g. `Topic.Attributes.SetValue()`, `Topic.References.SetTopic()`, or `Topic.Relationships.SetTopic()`. Those calls don't need to be updated just yet. Instead, we just need to make sure that a) the core attributes (such as `Key`, `ContentType`, and `ParentId`) are addressed, and check for duplicates for `VersionHistory` (which will be an issue when reloading an existing version).

While I was at it, I removed a legacy line of code which set the `ParentID` attribute to `IsDirty = false`. As the `ParentID` is no longer stored as an attribute, this did nothing outside of create an arbitrary attribute reference.