Introduced `markDirty` parameter to `SqlDataReader` extensions
Originally, the `SqlDataReader` extensions were exclusively intended to support loading the latest version of existing topics, in which case the topics (and their attributes, relationships, and topic references) should all be set to `markDirty=false`, since they are consistent with the values in the database.

By introducing support for `Load(referenceTopic, version)`, we break that assumption by allowing older versions of the topic to be loaded from the database. As such, the `SqlDataReader` needs to be updated to conditionally allow collections or values to be marked `IsDirty`.

This is acheived by adding in the nullable `markDirty` parameter. If ommitted, the default `IsDirty` handling will be applied (i.e., new, removed, or changed values will flag `IsDirty`). If it is set to `false`, then an update will not alter the `IsDirty` status of the `Topic` (i.e., it will return to whatever state it was at prior to the update—which could still be `IsDirty`).