Removed legacy version check
The `Version` column was added to the `GetTopics` stored procedure's `Attributes` query after OnTopic 4.0.0 was released. To maintain backward compatibility, a check was put in place to check the field count of the results, and only set the version if it was available.

With the upcoming release of 5.0.0, we can accept this as a breaking change—and, indeed, other aspects of the `SqlTopicRepository` won't work with the legacy 4.0.0 schema anyway, since it didn't support e.g. `TopicReferences`. As such, there's no further necessity or benefit for this.