Fixed bug in `GetTopicVersion` so it correctly returns core attributes
When the core attributes were moved from `Attributes` to `Topics`, the `GetTopics` stored procedure was correctly updated to look for the core attributes in `Topics`, instead of relying on the complex logic for extracting the latest version of them from `Attributes`. During this, however, `GetTopicVersion` was missed, and thus it continued to rely on the legacy logic.

Since the core attributes are no longer stored in the `Attributes` table, that means the core attributes were not returned—and, thus, `GetTopicVersion` returned invalid results.

This update fixes that issue by implementing the (much simpler!) query against the `Topics` table.