Fixed bug in which topic is returned from `LoadTopicGraph()``
When the `referenceTopic` was incorporated into `LoadTopicGraph()` (72f0560) a bug was inadvertantly introduced. `LoadTopicGraph()` is expected to return the topic that's being loaded, which will be the first `topicId` returned in the first data set, which represents the core topic attributes. Previously, this would _also_ be the first topic in the `topics` cache. With `referenceTopic`, however, the first topic in the `topics` cache will always be the root topic, regardless of which topic is being loaded.

For many scenarios, this is fine, as we usually load the entire topic cache at once. When loading an individual topic, however, such as the `Root:Configuration`, or a specific version via `Load(referenceTopic, version)`, this introduces a problem. This becomes a bug when the caller then operates against the result of `LoadTopicGraph()` or `Load()` expecting it to be the specific topic requested.

To remedy this, a check is injected into the loop over the first data set in order to grab the first `TopicID` and store it as a local `rootTopicId`, and then return the `Topic` associated with that at the end of `LoadTopicGraph()`.