Delete orphaned attributes caused by `Load(referenceTopic, version)`
When loading a specific version of a topic with an existing `referenceTopic`, all attribute values on the `referenceTopic` are overwritten with the values from the version. If, however, new attributes were introduced after that version, they won't be overwritten. That's because the data for the original version won't have a corresponding attribute key or value that's being set.

In these cases, we just need to determine if there are any newer attributes—i.e., attributes with a `LastModified` date that's newer than the `version`—and, if there are, remove them.