Introduced versioning support to the `Relationships` table
In addition to the `Version` column being added to both the table definition as well as the primary key constraint, the `Relationships` table also necessitates the introduction of an `IsDeleted` column.

This is because, unlike `Attributes` and `TopicReferences`, there can be more than one value (`Target_TopicID`, in this case) per key. As such, we can't just overwrite the value of a key by marking it as `null`. Instead, we must introduce an additional column to track the state of the record.