Introduced `Relationships`, `TopicReferences` versioning to `GetTopicVersion`
Previously, the `GetTopicVersion` stored procedure had no version support for topic references or relationships, and simply returned the `TopicReferences` and `Relationships` data for the given `@TopicID`. Now, it returns the latest version for the `TopicReferences` up to and including the provided `@Version`. This is a more complex query, but allows topic references and relationships to be restored to a previous state.