Introduced `Relationships`, `TopicReferences` versioning to `GetTopics`
Previously, the `GetTopics` stored procedure had no version support for topic references or relationships, and simply returned the `TopicReferences` and `Relationships` data for the given `@TopicID`. Now, it returns the latest version for the `TopicReferences`. This is a more complex query, but allows topic references and relationships to be versioned, similar to attributes.

While I was at it, I applied more consistent syntax for table aliases (always using the `AS` keyword, whereas previously this was inconsistently applied).