Added `IsDeleted` support to `SetRelationships()` extension
With the introduction of relationship versioning, the `SqlDataReader.SetRelationships()` extension method must account for the `IsDeleted` column; if it is set to true (`1`), then the relationship should be _deleted_ (if it exists), not added.

This ensures support for updating relationships on existing topics, which will be needed for e.g. the proposed `ITopicRepository.Refresh()` method, which will handle cache refreshes.

As part of this, I added a new `GetBoolean(columnName)` private helper method, to complement existing helpers such as `GetString(columnName)` and `GetInteger(columnName)`.