Updated `TopicReferences` to handle `NULL` columns
If a `TopicReference` is deleted, a new version should be created with a `NULL` value.

To support this, the `TopicReferences.Target_TopicID` column must be nullable, the `UpdateReferences` stored procedure must insert a `NULL` when `@DeleteUnmatched`, and, finally, `SqlDataReader.SetReferences()` must account for the potentially null return type.

This includes the addition of a new `GetNullableTopicId()` helper method, which returns a null `int?` if the response is `DBNULL`.