Preemptively clear `Relationships`, `References` during `Rollback()`
When doing a `Rollback()`—or, more specifically, a `Load(referenceTopic, version)` where the `topicId` is in the `referenceTopic` graph—we should preemptively clear the `Relationships` and `References` so that they can be reloaded with the target `version`.

Otherwise, since we don't store the `version` metadata for these relationships, there isn't a(n easy) way to detect if any relationships or referenced topics should be deleted.

Specifically, this occurred when those topics had been added for the first time after the loaded version—otherwise, we'd be able to deleted them due to an `IsDeleted` relationship, or `NULL` `Target_Topic` reference.