Implement basic unit test for the `UpdateTopic` stored procedure
The `UpdateTopicTest` creates a new topic, then updates the `Key` and `ContentType` using the `UpdateTopic` stored procedure. It confirms that the `Key` and `ContentType` are updated.

Note: Some of this was inadvertantly committed as part of the previous commit, which was supposed to be exclusive to the `UpdateRelationshipsTest` (04eb1bc). Apologies to future readers for the sloppy history.