Set SQL unit tests to compile during `debug`—but not `release`
The local SQL server won't be available on Azure DevOps. As such, we don't want the SQL server unit tests project to compile—and, thus, have its unit tests discovered—when running as part of continuous integration. We do want updates to the unit tests to be automatically compiled when running the SQL unit tests in Visual Studio, however.

Since Visual Studio's unit tests run in the `debug` configuration and Azure DevOps' run in the `release` configuration, we can hack around this by limiting the compilation of the unit tests project to the `debug` configuration.