Fixed bug in `GetChildTopicIDs` function
The unit test for the `GetChildTopicIDs` function (4c80bb0) exposed that it was not correctly behaving, as it was still written to query the `Attributes` table, even though `ParentID` had been moved to `Topics` along with other core, non-versioned attributes. This fixes the implemention of `GetChildTopicIDs`. (Though, it could be argued, that with the current implementation having a function might not provide much benefit.)