Implement basic unit test for the `GetTopicID` function
The `GetTopicID` attempts to load an existing topic using the `GetTopicID` function with a fully-qualified unique key. It confirms that the correct topic is returned.

As part of this, ensured that a `Root` topic with a `TopicID` of `1` is first established, as this is a dependency of `GetTopicID`. (We may want to revisit this dependency, but it remains a current dependency nevertheless.)