Ensured implicit `Root` topic (with `TopicID` of `1`)
Before executing the stored procedure tests, ensure there's a root topic with a `TopicID` set to `1`.

Most tests don't depend on this, but the `CreateTopic` stored procedure doesn't (yet) handle parentless topics well (in that it will assign them to overlapping ranges), and some functions (such as `GetUniqueKey`) still expect a topic with a `TopicKey` set to `Root` and a `TopicID` set to `1`.