Treat `TokenizedTopicList` as a `Relationship` or a `Reference`
Previously, the `TokenizedTopicList` defaulted to being treated as an attribute, unless it was explicitly set to `SaveAsRelationship`. Instead, it should _always_ default to a relationship, unless the `TokenLimit` is `1`, in which case it should be saved as a topic reference. Pointers to topics shoud never be saved as an attribute, as that prevents any sort of referential integrity or reciprocal relationship.