Migrated `TopicFactory` to use `DynamicTopicLookupService`
The `DynamicTopicLookupService` was originally intended for use with the `TopicFactory`, but we ended up using the `DefaultTopicLookupService` instead since we were moving away from `Topic` derivatives, and didn't expect to be creating any more. As such, dynamically assessing them was a bit of overhead considering we only expected `Topic`, `ContentTypeDescriptor`, and `AttributeDescriptor`.

That changes a bit with the introduction of strongly typed attribute descriptors (e.g., `TextAttribute`). And as we plan on migrating these to the OnTopic Editor, where they make more logical sense, we need to make sure that the `TopicFactory` can discover these, ideally, without needing to be manually registered.

Hopefully, this will allow us to optionally move attribute types to their own projects or assemblies.