Remove Topic Editor attribute types
Ideally, the attribute type descriptors should be registered alongside their corresponding attributes in the OnTopic Editor. By having them in the OnTopic Library, we create an awkward dependency between the Editor and core library which makes maintenance difficult.

In the previous commit, we migrated `TopicFactory` to use the `DynamicTopicLookupService` (b0afb90) with the expectation that it will then be able to find these in the OnTopic Editor, assuming it's in scope. In a corresponding update, we'll add these back as part of the OnTopic Editor itself.

As part of this, five of these—`BooleanAttribute`, `NestedTopicListAttribute`, `RelationshipAttribute`, `TextAttribute`, `TopicReferenceAttribute`—have been moved to `OnTopic.TestDoubles` so that the unit tests that use them as examples can continue to operate. As these are bare-bones classes, this duplication isn't considered much of a concern.

As part of this, namespaces in unit tests needed to be updated to point to `OnTopic.TestDoubles.Metadata` where appropriate.