Moved suppression of CA1711
Normally, in .NET, classes should only end in `Attribute` if they inherit from the `Attribute` class. In this case, the naming conventions of OnTopic Editor conflict with the naming conventions of .NET. That's unfortunate, but well-established. As such, we're suppressing this warning since this is a deliberate design decision.

This suppression was originally defined on OnTopic, then moved to OnTopic Tests. With the attribute type descriptors finally moved to Test Doubles (a380d90), the suppression should follow.