Introduce `IsLoaded` property to `TopicReferenceDictionary`
The `IsLoaded` property allows us to track if a topic was not able to be fully loaded during `ITopicRepository.Load()` due to the `Target_TopicId` not existing in the current topic graph. This generally occurs because a) the load is limited to an individual topic or branch, and b) the `referenceTopic` was not passed to `Load()`, or itself represents an incomplete topic graph.