Added in basic support for cache updates
This approach isn't perfect in that it prevents the request from being completed until the refresh is finished, instead of being performed on a background thread. A more sophisticated version would operate as a background task. Alternatively, in the future, we'll likely introduce a queue via e.g. Azure Service Bus to track events, and rely on those to trigger `Refresh()`. Until then, however, this provides a simple approach for satisfying pulling in occassional cache updates.