Introduced new `InvalidTypeException`
The `InvalidTypeException` is thrown is the `TopicMappingService` attempts to map to a type (such as a `TopicViewModel`) which cannot be located in the associated `ITypeLookupService`. This can happen, for instance, if the topic to be mapped has a content type that doesn't have a corresponding `TopicViewModel`, or if that type isn't properly registered with the `ITypeLookupService`.

This inherits from the existing `TopicMappingException`, thus allowing callers to catch the general exception, or hone in on this specific exception if appropriate.