Introduced new `MappingModelValidationException`
The `MappingModelValidationException` is thrown a mapping service attempts to map to a model which cannot be mapped to. This is especially important with the `ReverseTopicMappingService`, since it takes data integrity very seriously when writing back to the database and, thus, fails if any members don't align with the `ContentTypeDescriptor`.

By contrast, this won't generally be used with the read-only mapping services, such as `TopicMappingService`, since it is generally designed to be forgiving.

This inherits from the existing `TopicMappingException`, thus allowing callers to catch the general exception, or hone in on this specific exception if appropriate.