Migrated topic view and binding models to records
The new C# 9.0 record feature is optimal for view models and binding models, which we don't expect to change after they've been generated. Records are also potentially a bit faster, since they're immutable, and help ensure that they're read-only. Finally, the record syntax makes it easier to generate new versions should we need to make modifications via the `with` syntax.

The nature of records mandates that all downstream implementations of view and binding models are also converted to records, so this is (very much!) a breaking change.