Unobsoleted `ITopicRepository` events
Previously, the `DeleteEvent`, `MoveEvent`, and `RenameEvent`, as well as their respective `EventArgs` classes, had all been marked as deprecated. That was premature. While, admittedly, they aren't currently in use, they continue to provide potential value, and don't introduce much overhead to maintain. We can reevaluate these in the future, if we wish, but for now they'll be offerred a reprieve.