Generalize `SqlDataReader` extensions to operate off `IDataReader`
The `SqlDataReader` can only be created from a `SqlConnection` object, and cannot be separately constructed. That makes it difficult to write unit tests against. By generalizing the `SqlDataReader` extensions to actually target the underlying `IDataReader`, it still works successfully against the target `SqlDataReader`, but _also_ works against e.g. the `DataTableReader` class, which can easily be constructed for the purpose of unit tests.

There is one method—the `SetExtendedAttributes()` method—which cannot be successfully migrated to use `IDataReader`. That's because it relies on `GetSqlXml()`, which is specific to `SqlDataReader`. Given that, I've added a wrapper around that call so that it's only called if the `IDataReader` can be cast as a `SqlDataReader`. In practice, we expact that to be in all cases _except_ for the unit tests.

Because the class is still _intended_ to target the `SqlDataReader`, and because it still contains at least one method that is specific to `SqlDataReader`, I'm keeping the name as `SqlDataReaderExtensions`, instead of renaming it to e.g. `DataReaderExtensions`. We may reevaluate that in the future—though it doesn't matter _too_ much since it's an internal class.