Established `AttributesDataTable` for unit testing
While we can't easily test `SqlDataReader`, we can easily test a `DataTableReader`, which operates off of the same `IDataReader` interface as `SqlDataReader`. Given this, the `AttributesDataTable` provides a simple way of mimicking SQL response data for the `IDataReader` extension methods used by `SqlTopicRepository` to operate off of, thus allowing us to unit test these methods without using a SQL database.

The `AttributesDataTable` class includes not only the basic schema definition for the `Attributes` table, but also an `AddRow()` method with strongly typed parameters to simplify creating a new record.