Established `ExtendedAttributesDataTable` for unit testing
While we can't easily test `SqlDataReader`, we can easily test a `DataTableReader`, which operates off of the same `IDataReader` interface as `SqlDataReader`. Given this, the `ExtendedAttributesDataTable` provides a simple way of mimicking SQL response data for the `IDataReader` extension methods used by `SqlTopicRepository` to operate off of, thus allowing us to unit test these methods without using a SQL database.

The `ExtendedAttributesDataTable` class includes not only the basic schema definition for the `ExtendedAttributes` table, but also an `AddRow()` method with strongly typed parameters to simplify creating a new record.

Note: Because the `SetExtendedAttributes()` extension method relies upon the `GetSqlXml()` method, which is not available on the `IDataReader` interface, we may not be able to actually unit test the extended attributes using the same technique. That said, we're creating this data table schema in case we can identify a way of mitigating that.