Introduced unit test for validating the `SetReferences()` extension
This unit test creates a `TopicsDataTable` and a `TopicReferencesDataTable` and passes them to the `LoadTopicGraph()` extension method, ensuring that the (private) `SetReferences()` extension method correctly utilizes the data tables to set the specified topic reference on the new `Topic` entity. While it's at it, it also confirms that the collection is correctly marked as `IsDirty()`, since it has changed.