Introduced unit test for validating `referenceTopic` behavior
This unit test creates a `TopicsDataTable` and a `TopicReferencesDataTable` and passes them to the `LoadTopicGraph()` extension method, ensuring that the (private) `SetReferences()` extension method correctly utilizes the data tables along with a `referenceTopic` to set the reference to the existing `Topic` on the new `Topic` entity. While it's at it, it also confirms that the collection is correctly marked as `IsDirty()`, since it has changed, as well as `IsFullyLoaded`, since all references should be resolved.