Removed `IsExternalInit` polyfills
The .NET Core 3.x SDK didn't include the `IsExternalInit` class needed to enable C# 9.0's `init` property initializers, which are especially useful on C# 9.0's `record` types, although can also be used elsewhere. Now that we have upgraded to .NET 5.0 (713051c), this polyfill is no longer needed, and can be deleted. (And, in fact, keeping it actually breaks the implementation.)