CS8601: Possible null reference assignment
In practice, the values in the underlying `_storage` dictionary will never be `null`, as they are validated on entry, and deleted if the values are set to `null`. As such, we can be confident that if `TryGetValue()` is successful, then the output value will not be `null`.

Note: This is required to satisfy the expectations of `IDictionary<TKey, TValue>`. Given that, it's peculiar that this throws a CS8601.