CS8603: Possible null reference return
In each of these cases, the return values _could_ potentially be null, and callers should be aware of that. In most of these cases, which were in `TopicMappingServiceTest`, the caller was an `Assert.IsNotNull()`, and was thus already _expecting_ a potentially null reference—otherwise there'd be no point in testing for it.