CS8602: Dereference of a possibly null reference
I previously patched several CS8602s (2d426fb ), but the fix of CS8603 (cfa8ac9) introduced a new one. This possibility was already accounted for in the return condition, but not properly accounted for prior to introducing new code for deleting unmatched attributes. Given that, I've moved the validation up from the return condition to above the orphaned attribute handling, so it addresses both scenarios.