CS8604: Possible null reference argument for parameter
Where possible, a default was provided using the null-coalescing operator (`??`).  In most cases, however, I had to add a `Contract.Assume()` or manually check and throw an exception. In one case, I needed to suppress the warning, unfortunately, as otherwise we'll need to rewrite that logic to account for the scenario. (In practice, the target accepts nulls, despite the warning.)