Removed suppression of CS8604
This partially rolls back one of the changes made when addressing CS8604 (c3ed628), which is the suppression of a warning introduced by improved nullability annotations in .NET 5.0. Now that we're targeting .NET Core 3.1, this suppression is no longer relevant. We can revisit at the point that we upgrade OnTopic.