Renamed attribute types descriptors to include `Descriptor` suffix
Previously, the content types for attribute types, such as `TextAttribute`, ended with the `Attribute` suffix. This introduces two points of confusion. First, they are actually attribute _descriptors_—and, indeed, derive from `AttributeDescriptor`. Second, in their C# implementation, this causes problems because the `Attribute` suffix is usually reserved for actual `[Attribute]`s.

To remedy these issues, the content type is renamed to use the `AttributeDescriptor` suffix (e.g., `TextAttributeDescriptor`). This requires updating the
keys of the appropriate `ContentTypeDescriptor`s, as well as the content types of the corresponding implementations of those attribute descriptors in the database.