Strip `AttributeDescriptor` from `EditorType` value
The `EditorType` is intended to enforce the convention of converting the current type name—representing a content type—to the identifier used by the OnTopic Editor. Previously, this involved stripping the `Attribute` suffix from the type name. Now that we use the more descriptive and accurate `AttributeDescriptor` suffix (9ba3359), this needs to be updated to account for that.